moves API types to /api folder
