+++ app/layout/projects/new/form/constants.ts
@@ -1,8 +1,8 @@
-import { PlanningUnit } from 'types/project-model';
+import { PlanningArea } from 'types/api/project';
 
-export const DEFAULT_AREA = {
+export const DEFAULT_AREA: Omit<PlanningArea, 'country'> = {
   planningUnitAreakm2: 10,
-  planningUnitGridShape: PlanningUnit.HEXAGON,
+  planningUnitGridShape: 'hexagon',
 };
 
 export const PA_OPTIONS = [
