--- app/layout/projects/new/form/planning-area-selector/planning-unit-grid/planning-unit-button/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
