--- app/layout/projects/new/form/planning-area-selector/planning-unit-grid/types.ts
@@ -1,6 +1,0 @@
-import { PlanningUnit } from 'types/project-model';
-
-export interface PlanningUnitGridProps {
-  unit: PlanningUnit;
-  onChange?: (value: PlanningUnit) => void;
-}
