+++ app/store/slices/scenarios/detail.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { injectReducer } from 'store';
-import { Solution } from 'types/project-model';
+import { Solution } from 'types/api/project';
 import { ScenarioSidebarTabs } from 'utils/tabs';
 
 interface ScenarioShowStateProps {
