+++ app/store/slices/scenarios/edit.ts
@@ -3,7 +3,7 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import type { ScenarioPlanningUnit } from 'hooks/scenarios/types';
 
 import { injectReducer } from 'store';
-import { Solution } from 'types/project-model';
+import { Solution } from 'types/api/project';
 import { ScenarioSidebarTabs } from 'utils/tabs';
 
 import type { PUAction } from './types';
